+++ gpMgmt/bin/ggrebalance
@@ -14,7 +14,7 @@ try:
     from gppylib.system.environment import GpCoordinatorEnvironment
     from gppylib.system import configurationInterface, configurationImplGpdb
     from gprebalance_modules.shrink import GGShrink, DBNAME
-    from gprebalance_modules.ggrebalanceSM import RebalanceSM, GGRebalanceSM
+    from gprebalance_modules.ggrebalance_main_sm import GGRebalanceMainSM
     from gprebalance_modules.planner import *
 except ImportError as e:
     sys.exit('ERROR: Cannot import modules.  Please check that you have sourced greenplum_path.sh.  Detail: ' + str(e))
@@ -283,7 +283,7 @@ def main(options, args, parser):
             sys.exit(1)
 
         global gg_main_rebalance_SM
-        gg_main_rebalance_SM = GGRebalanceSM(logger, dburl, options, gpenv, gparray, gparray_dump_filename)
+        gg_main_rebalance_SM = GGRebalanceMainSM(logger, dburl, options, gpenv, gparray, gparray_dump_filename)
         gg_main_rebalance_SM.run()
 
         sys.exit(0)
