+++ gpMgmt/bin/gprebalance_modules/Makefile
@@ -11,7 +11,8 @@ rebalance_status.py \
 shrink.py planner.py \
 rebalance_commons.py \
 rebalance_schema.py \
-solver.py ggrebalanceSM.py
+solver.py ggrebalance_sm.py \
+ggrebalance_main_sm.py
 
 installdirs:
 	$(MKDIR_P) '$(DESTDIR)$(bindir)/gprebalance_modules'
